app/requirements.txt
# FILE: app/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sentence-transformers
faiss-cpu
numpy
motor
python-dotenv
setuptools
cachetools